added image_service class
A class which handles the saving of images which was previously on the fda_search class. This is so that we can reuse the same code whenever we want to retake the image of a medication. Note that this class is using the bad practice of buildcontexts in third party classes. We will want to refactor this in the future, but while this is a small project this should be Okay.